Various small patches
 - Remove -pedantic from compile_flags.txt (it's annoying and
   unnecessary)
 - Change the current time (in milliseconds) tracked by the PIT from a
   32-bit integer to a 64-bit integer.
   Theoretically slows down the code slightly (we're compiling for
   32-bit so software support is needed for 64-bit integers), but
   changes the maximum time from ~50 days to ~600 million years. This
   might be overkill, since I don't intend on running the kernel for
   more tan perhaps a few hours at a time, but we'll see if I change it
   back later
 - Add backspace support to the VGA driver (a bit half-arsed at the
   moment...)
 - Remove leftover debug print statements from the IDT initialisation
   code
 - Remove code from kernel_main which was just there to demo the fact
   that I got C++ features (like constructors and destructors) working +
   code to demonstrate scrolling works
 - Added some short delay using io_wait to show that the PIT is still
   correctly counting that time has passed (it kept consistently showing
   me 3ms before, now it's around 180~200)

+++ include/pit.hpp
@@ -29,7 +29,7 @@
 extern "C" void pit_handle_trigger(void);
 
 namespace pit {
-	volatile extern uint32_t millis;
+	volatile extern uint64_t millis;
 
 	void init_pit0(); // initialise channel 0 to cause an IRQ0 every millisecond
